refactor(tool): move version metadata to util package
